+++ build.sh
@@ -1,4 +1,30 @@
 #!/bin/bash -e
+
+if [ "${BUILD_TYPE}" != "assemble" || "${BUILD_TYPE}" != "bundle" ]; then
+    echo "Invalid build-type: " ${BUILD_TYPE}
+    exit 1
+fi
+
+if [ "${UPLOAD_TO_PLAY_STORE}" == true ]; then
+    if [ -n "${PACKAGE_NAME}" ]; then
+        echo "package-name is required when uploading to play store"
+        exit 2
+    elif [ -n "${KEYSTORE_CONTENT}" ]; then
+        echo "Missing keystore-content"
+        exit 3
+    elif [ -n "${KEYSTORE_PASSWORD}" ]; then
+        echo "Missing keystore-password"
+        exit 4
+    elif [ -n "${KEYSTORE_ALIAS}" ]; then
+        echo "Missing keystore-alias"
+        exit 5
+    elif [ -n "${JSON_KEY_DATA}" ]; then
+        echo "Missing json-key-data"
+        exit 6
+    fi
+fi
+
+export PLAY_STORE_JSON_KEY_DATA=${JSON_KEY_DATA}
 OUTPUT_PATH="${PWD}/${OUTPUT_PATH}"
 
 # If the variable KEYSTORE_CONTENT is set, then the keystore is created from the content of the variable.
@@ -7,12 +33,25 @@ if [ -n "${KEYSTORE_CONTENT}" ]; then
     echo $KEYSTORE_CONTENT | base64 --decode > "keystore.jks"
 fi
 
-sudo gem install bundler -NV
+# If the variable BUNDLER_VERSION is set, then install bundler with selected version
+# otherwise use latest
+if [ -n "${BUNDLER_VERSION}" ]; then
+    echo "Running bundler with version: ${BUNDLER_VERSION}"
+    gem install bundler:${BUNDLER_VERSION} -NV
+else
+    gem install bundler -NV
+fi
 
-script_path=$(cd $(dirname ${0}); pwd)
-cp -r ${script_path}/fastlane ./
-cp -r ${script_path}/Gemfile ./
+cp -r ${ACTION_PATH}/fastlane ./
+cp -r ${ACTION_PATH}/Gemfile ./
 
 bundle install
 
-fastlane android $BUILD_TYPE
+# If the variable FASTLANE_ENV is set then run fastlane with the --env equal to the variable.
+if [ -n "${FASTLANE_ENV}" ]; then
+    echo "Running fastlane with environment: ${FASTLANE_ENV}"
+    fastlane --env ${FASTLANE_ENV} android $BUILD_TYPE
+else
+    echo "Running fastlane"
+    fastlane android $BUILD_TYPE
+fi